requirements.txt
# FastAPI 백엔드
fastapi==0.95.2
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==1.10.13

# 데이터베이스
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
alembic==1.13.0

# 벡터 데이터베이스
weaviate-client==3.25.3

# 비동기 처리
celery==5.3.4
redis==5.0.1

# AI/ML 모델 (temporarily commented out)
# openai==1.3.0
# transformers==4.35.0
# torch>=2.1.0
# sentence-transformers==2.2.2
# opencv-python==4.8.1.78
# pillow>=10.0.0
# numpy>=1.24.0

# 유틸리티
python-dotenv==1.0.0
httpx==0.25.2
